On the 32 bit systems the type long is always 32 bit, and it is not enough in this case, so registered services has never called

+++ src/rvi.c
@@ -1820,7 +1820,7 @@ int rviReadRcv( TRviHandle handle, json_t *msg, TRviRemote *remote )
     tmp = json_object_get( msg, "data" );
     if( !tmp ) { err = RVI_ERR_JSON; goto exit; }
 
-    long timeout = json_integer_value( json_object_get( tmp, "timeout" ) );
+    long long timeout = json_integer_value( json_object_get( tmp, "timeout" ) );
     time(&rawtime);
     if( rawtime > timeout ) { err = RVI_ERR_JSON; goto exit; }
